requirements.txt
absl-py==1.4.0
appdirs==1.4.4
astunparse==1.6.3
biopython==1.80
biotite==0.35.0
brotlipy==0.7.0
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
contourpy==1.0.7
crc32c==2.3.post0
cryptography @ file:///croot/cryptography_1673298753778/work
cycler==0.11.0
filelock==3.9.0
flatbuffers==23.1.21
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
gast==0.4.0
gemmi==0.5.8
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.8.0
huggingface-hub==0.12.0
idna @ file:///croot/idna_1666125576474/work
importlib-metadata==6.0.0
interchange==2021.0.4
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
lightgbm @ file:///croot/lightgbm_1677766491914/work
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
mock==5.0.1
monotonic==1.6
msgpack==1.0.4
networkx==3.0
numpy @ file:///croot/numpy_and_numpy_base_1672336185480/work
oauthlib==3.2.2
obonet==0.3.1
opt-einsum==3.3.0
packaging @ file:///croot/packaging_1678965309396/work
pandas==1.5.3
pansi==2020.7.3
pickle5==0.0.11
Pillow==9.3.0
pooch @ file:///tmp/build/80754af9/pooch_1623324770023/work
protobuf==3.19.6
py2neo==2021.2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.14.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
subword-nmt==0.3.8
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
tfrecord==1.14.1
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.1
transformers==4.26.0
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
urllib3 @ file:///croot/urllib3_1673575502006/work
Werkzeug==2.2.2
wget==3.2
wrapt==1.14.1
xgboost==1.7.3
zipp==3.12.0
